requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
requests
APScheduler
python-telegram-bot
mysqlclient
